requirements.txt
python-telegram-bot[rate-limiter]==21.7
pandas==2.2.3
openpyxl==3.1.5
XlsxWriter==3.2.0
beautifulsoup4==4.12.3
lxml==5.3.0
